+++ src/main/java/com/yor42/projectazure/gameobject/entity/ai/tasks/CompanionSteerEntityTask.java
@@ -104,7 +104,7 @@ public class CompanionSteerEntityTask extends Task<AbstractEntityCompanion> {
                     float f = 0.04F;
                     float oldyaw = boat.yRot;
                     boat.yRot = (float) (MathHelper.atan2(d1, d0) * (double) (180F / (float) Math.PI)) + 90.0F;
-                    boat.setDeltaMovement(boat.getDeltaMovement().add(MathHelper.sin(-boat.yRot * ((float) Math.PI / 180F)) * f, 0.0D, (double) (MathHelper.cos(boat.yRot * ((float) Math.PI / 180F)) * f)));
+                    boat.setDeltaMovement(boat.getDeltaMovement().add(MathHelper.sin(-boat.yRot * ((float) Math.PI / 180F)) * f, 0.0D, MathHelper.cos(boat.yRot * ((float) Math.PI / 180F)) * f));
                     float newyaw = boat.yRot;
                     float x1 = (float) Math.sqrt(boat.getDeltaMovement().x()*boat.getDeltaMovement().x()+boat.getDeltaMovement().z()*boat.getDeltaMovement().z());
                     boat.setPaddleState(oldyaw - newyaw < -0 || x1>0.2, oldyaw - newyaw > 0 || x1>0.2);
