+++ src/main/java/com/yor42/projectazure/gameobject/entity/projectiles/EntitySpellBall.java
@@ -49,11 +49,11 @@ public class EntitySpellBall extends DamagingProjectileEntity {
     }
 
     public void shoot(double x, double y, double z, float velocity, float inaccuracy) {
-        Vector3d vector3d = (new Vector3d(x, y, z)).normalize().add(this.random.nextGaussian() * (double)0.0075F * (double)inaccuracy, this.random.nextGaussian() * (double)0.0075F * (double)inaccuracy, this.random.nextGaussian() * (double)0.0075F * (double)inaccuracy).scale((double)velocity);
+        Vector3d vector3d = (new Vector3d(x, y, z)).normalize().add(this.random.nextGaussian() * (double)0.0075F * (double)inaccuracy, this.random.nextGaussian() * (double)0.0075F * (double)inaccuracy, this.random.nextGaussian() * (double)0.0075F * (double)inaccuracy).scale(velocity);
         this.setDeltaMovement(vector3d);
         float f = MathHelper.sqrt(getHorizontalDistanceSqr(vector3d));
         this.yRot = (float)(MathHelper.atan2(vector3d.x, vector3d.z) * (double)(180F / (float)Math.PI));
-        this.xRot = (float)(MathHelper.atan2(vector3d.y, (double)f) * (double)(180F / (float)Math.PI));
+        this.xRot = (float)(MathHelper.atan2(vector3d.y, f) * (double)(180F / (float)Math.PI));
         this.yRotO = this.yRot;
         this.xRotO = this.xRot;
     }
@@ -100,7 +100,7 @@ public class EntitySpellBall extends DamagingProjectileEntity {
     public boolean hurt(DamageSource p_70097_1_, float p_70097_2_) {
         if(p_70097_2_>=8) {
             for (int i2 = 0; i2 < 8; ++i2) {
-                this.level.addParticle(ParticleTypes.PORTAL, (double) this.getX() + random.nextDouble(), (double) this.getY() + 1.2D, (double) this.getZ() + random.nextDouble(), 0.0D, 0.0D, 0.0D);
+                this.level.addParticle(ParticleTypes.PORTAL, this.getX() + random.nextDouble(), this.getY() + 1.2D, this.getZ() + random.nextDouble(), 0.0D, 0.0D, 0.0D);
             }
             this.remove();
         }
